+++ src/scroll-position/constants.js
@@ -1,1 +1,3 @@
+// This constant is imported in various files so it was "extracted" here
+// to a separate file in order to not cause any "circular dependency" issues.
 export const PAGE_SCROLLABLE_CONTAINER_KEY = 'page';